requirements.txt
transformers>=4.30.0
torch>=2.0.1
numpy>=1.24.3
pandas>=2.0.2
scikit-learn>=1.2.2
flask>=2.3.2
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
openai>=1.3.0
nltk>=3.8.1
python-multipart>=0.0.6
pytest>=8.4.2